+++ action_app/lib/analyze_command.dart
@@ -1,5 +1,5 @@
 import 'package:actions_toolkit_dart/core.dart';
-import 'package:dart_code_metrics/lint_analyzer.dart';
+import 'package:flutterando_metrics/lint_analyzer.dart';
 import 'package:github/github.dart';
 
 import 'arguments.dart';