+++ action_app/test/github_workflow_utils_test.dart
@@ -81,7 +81,7 @@ void main() {
         throwsArgumentError,
       );
 
-      const slug = 'dart-code-checker/dart-code-metrics-action';
+      const slug = 'dart-code-checker/flutterando-metrics-action';
 
       expect(
         const GitHubWorkflowUtils(
@@ -111,7 +111,7 @@ void main() {
     });
 
     test('isTestMode returns true only for current repo', () {
-      const slug = 'dart-code-checker/dart-code-metrics-action';
+      const slug = 'dart-code-checker/flutterando-metrics-action';
 
       expect(
         const GitHubWorkflowUtils(