+++ action_app/test/report_utils_test.dart
@@ -1,5 +1,5 @@
 import 'package:action_app/report_utils.dart';
-import 'package:dart_code_metrics/lint_analyzer.dart';
+import 'package:flutterando_metrics/lint_analyzer.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';