+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## 3.0.0
 
-* feat!: Support Dart Code Metrics 3.x.x
+* feat!: Support Flutterando Metrics 3.x.x
 
 ## 2.0.0